requirements.txt
Flask
pymongo
python-dotenv
minio
openai-whisper
